:lock: 导入设置前进行备份

+++ Plain Craft Launcher 2/Pages/PageSetup/ModSetup.vb
@@ -702,12 +702,12 @@ Public Class ModSetup
         Try
             Log($"[Setup] 导出设置：到 {Target}，{If(Registry, "含注册表", "不含注册表")}")
             If File.Exists(Target) Then File.Delete(Target) '选文件的时候已经确认要给他替换掉了
-            File.Copy(Path & "PCL\Setup.ini", Target)
+            CopyFile(Path & "PCL\Setup.ini", Target)
             If Registry Then
                 ShellAndGetOutput("cmd", Arguments:=$"/C reg export HKCU\Software\{RegFolder} ""{PathTemp}\ExportTemp.reg""")
                 WriteIni(Target, "RegInfo", Convert.ToBase64String(File.ReadAllBytes($"{PathTemp}\ExportTemp.reg")))
+                If File.Exists($"{PathTemp}\ExportTemp.reg") Then File.Delete($"{PathTemp}\ExportTemp.reg")
                 WriteIni(Target, "Identify", Setup.Get("Identify"))
-                File.Delete($"{PathTemp}\ExportTemp.reg")
             End If
             Return True
         Catch ex As Exception
@@ -723,6 +723,7 @@ Public Class ModSetup
     Public Function SetupImport(Source As String) As Byte
         Try
             Log($"[Setup] 导入设置：从 {Source}")
+            CopyFile(Path & "PCL\Setup.ini", Path & "PCL\Setup.ini.old")
             If ReadIni(Source, "Identify", "null") = "null" Then
                 Hint("不是有效的配置文件！", HintType.Critical)
                 Return 2 '失败
@@ -738,36 +739,38 @@ Public Class ModSetup
                 Dim key As String = pair.First()
                 Dim val As String = If(pair.Length < 2, "", pair.Last())
                 If Ln.StartsWithF("RegInfo:") Then '注册表
-                    File.WriteAllBytes($"{PathTemp}\ImportTemp.reg", Convert.FromBase64String(val))
+                    WriteFile($"{PathTemp}\ImportTemp.reg", Convert.FromBase64String(val))
                     hasReg = True
                 ElseIf Ln.StartsWithF("Identify:") Then
                     If val <> "" Then id = val
                 Else
+                    Log($"[Setup] 设置项：{key} 设置为 {val}")
                     Setup.Set(key, val, ForceReload:=True)
                 End If
             Next
-            If hasReg Then
-                If Setup.Get("Identify") <> id Then
+            If Not hasReg Then Return 1 '不重启
+            If Setup.Get("Identify") <> id Then
 #If BETA Then
-                    Dim msg As String = " PCL 隐藏主题、正版账号、LittleSkin 账号"
+                Dim msg As String = " PCL 隐藏主题、正版账号、LittleSkin 账号"
 #Else
-                    Dim msg As String = "正版账号、LittleSkin 账号"
+                Dim msg As String = "正版账号、LittleSkin 账号"
 #End If
-                    If MyMsgBox(
+                If MyMsgBox(
                         $"导入的设置可能来自另一台电脑，且将会导致{msg}等信息失效。" & vbCrLf &
                         "即使你导出了当前设置，也可能无法再次还原！" & vbCrLf & 'youzi-2333：别问我是怎么知道的
                         "除非你知道你在做什么，否则请取消导入注册表项！" & vbCrLf &
                         vbCrLf &
                         "点击 ""取消"" 按钮后，账号信息、主题颜色信息等不会被导入，其它安全的信息仍会导入。" & vbCrLf &
-                        "这是最后的警告！", Title:="警告",
+                        "这是最后的警告！" & vbCrLf &
+                        "如果你刚刚重装了系统，并在同一台电脑上启动 PCL，请忽略该提示并继续。", Title:="警告",
                         Button1:="取消", Button2:="我知道我在做什么！", Button3:="取消", IsWarn:=True) <> 2 Then
-                        Return 1 '不重启
-                    End If
+                    If File.Exists($"{PathTemp}\ExportTemp.reg") Then File.Delete($"{PathTemp}\ExportTemp.reg")
+                    Return 1 '不重启
                 End If
-                ShellAndGetOutput("cmd", Arguments:=$"/C reg import ""{PathTemp}\ImportTemp.reg""")
-                Return 0 '重启
             End If
-            Return 1 '不重启
+            ShellAndGetOutput("cmd", Arguments:=$"/C reg import ""{PathTemp}\ImportTemp.reg""")
+            If File.Exists($"{PathTemp}\ExportTemp.reg") Then File.Delete($"{PathTemp}\ExportTemp.reg")
+            Return 0 '重启
         Catch ex As Exception
             Log(ex, "导入设置失败", Level:=LogLevel.Msgbox)
             Return 2 '失败
